Lists the required packages

The `packages.yml` file contains a list of packages that are required.

pkg_list(path = ".", install = FALSE, scan = TRUE, rmd_dir = c("TD",
  "lectures", "site"))

Arguments

path
character vector pointing to the root of the bs2 repository containing the optional `packages.yml` file.

install
boolean telling whether to install missing packages (defaults to `FALSE`)

scan
boolean whether to scan source files for used packages (defaults to `TRUE`)

rmd_dir
character vector containing the subfolders with the Rmd source files.

Value

A tibble listing the package, the version, whether it is already installed or not.